requirements.txt
# Instagram Reels Microservice Requirements

# Core dependencies
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0

# Instagram scraping
instaloader==4.10
selenium==4.15.2
webdriver-manager==4.0.1

# AI and ML
sentence-transformers==2.2.2
torch==2.1.0

# Data processing
pandas==2.1.1
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
beautifulsoup4==4.12.2
tqdm==4.66.1
urllib3==1.26.18

# Development and testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# YouTube API integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Security updates (recommended)
# These versions include security patches
certifi>=2023.7.22  # Security patch for SSL certificates
